Move service configuration and error into module
- update references in module files
- update module name expectations in specs
