+++ modules/decision_reviews/spec/sidekiq/submit_upload_spec.rb
@@ -226,7 +226,7 @@ RSpec.describe DecisionReviews::SubmitUpload, type: :job do
                 expect do
                   subject.perform_async(appeal_submission_upload.id)
                   subject.drain
-                end.to raise_error(DecisionReviewV1::ServiceException)
+                end.to raise_error(DecisionReviews::V1::ServiceException)
                   .and trigger_statsd_increment('shared.sidekiq.default.DecisionReviews_SubmitUpload.enqueue', times: 1)
                   .and trigger_statsd_increment('shared.sidekiq.default.DecisionReviews_SubmitUpload.dequeue', times: 1)
                   .and trigger_statsd_increment('decision_review.form_10182.evidence_upload_retrieval.success',
@@ -252,7 +252,7 @@ RSpec.describe DecisionReviews::SubmitUpload, type: :job do
                 expect do
                   subject.perform_async(appeal_submission_upload.id)
                   subject.drain
-                end.to raise_error(DecisionReviewV1::ServiceException)
+                end.to raise_error(DecisionReviews::V1::ServiceException)
                   .and trigger_statsd_increment('shared.sidekiq.default.DecisionReviews_SubmitUpload.enqueue', times: 1)
                   .and trigger_statsd_increment('shared.sidekiq.default.DecisionReviews_SubmitUpload.dequeue', times: 1)
                   .and trigger_statsd_increment('decision_review.form_10182.evidence_upload_retrieval.success',
@@ -410,7 +410,7 @@ RSpec.describe DecisionReviews::SubmitUpload, type: :job do
                 expect do
                   subject.perform_async(appeal_submission_upload.id)
                   subject.drain
-                end.to raise_error(DecisionReviewV1::ServiceException)
+                end.to raise_error(DecisionReviews::V1::ServiceException)
                   .and trigger_statsd_increment('shared.sidekiq.default.DecisionReviews_SubmitUpload.enqueue', times: 1)
                   .and trigger_statsd_increment('shared.sidekiq.default.DecisionReviews_SubmitUpload.dequeue', times: 1)
                   .and trigger_statsd_increment('decision_review.form_995.evidence_upload_retrieval.success',
@@ -437,7 +437,7 @@ RSpec.describe DecisionReviews::SubmitUpload, type: :job do
                 expect do
                   subject.perform_async(appeal_submission_upload.id)
                   subject.drain
-                end.to raise_error(DecisionReviewV1::ServiceException)
+                end.to raise_error(DecisionReviews::V1::ServiceException)
                   .and trigger_statsd_increment('shared.sidekiq.default.DecisionReviews_SubmitUpload.enqueue', times: 1)
                   .and trigger_statsd_increment('shared.sidekiq.default.DecisionReviews_SubmitUpload.dequeue', times: 1)
                   .and trigger_statsd_increment('decision_review.form_995.evidence_upload_retrieval.success',
